Migrated to GoLang; more efficiency and scalability. Added new features after 4 years of hiatus caused my idleness, enhancing functionality and user experience
