+++ README.md
@@ -1,74 +1,60 @@
-# Kubenpod
+# KubenPod CLI Tool
 > Note that everything is experimental and may change significantly at any time.
 
-This tool aims to make it easier to list the pods that work in node on Kubernetes Cluster.
-This can be done with a node name or a pod name as you can see at examples below.
-If you encounter any problem or complexity, you can post it here as a comment.
+- KubenPod is a command-line interface tool for Kubernetes that aims to make it easier to list the pods that work in node on Kubernetes Cluster. This tool uses the Kubernetes and Metrics APIs to fetch data about pods and their resource usage.
+- This can be done with a node name or a pod name as you can see at examples below.
+If you encounter any problem or complexity, open an issue.
 
 ## Installation
+Before using the KubenPod CLI, ensure you have configured your Kubernetes context and have appropriate permissions to access the resources.
 ```
 sudo curl -LO https://raw.githubusercontent.com/engincanhoke/toppod/master/kubenpod > kubenpod;
 sudo chmod +x kubenpod;
 sudo mv kubenpod /usr/local/bin/kubenpod
 ```
+
 ## Usage
+The KubenPod CLI provides two main commands:
+1. `top` This command is used to show resource metrics (CPU and memory) of all pods on a specified node or the node of a specified pod.
+```bash
+kubenpod top [NODE_NAME]
+kubenpod top --pod-name [POD_NAME]
 ```
-$ kubenpod ip-10-0-0-1.us-east-1.compute.internal
+2. `list` This command lists all pods on a specified node or the node of a specified pod.
+```bash
+kubenpod list [NODE_NAME]
+kubenpod list --pod-name [POD_NAME]
+```
 
-hoke              hoke-deployment-cda628x58-p1vzz                                   176m   267Mi
-kube-system       aws-node-r6n1m                                                    3m     28Mi
-kube-system       coredns-165c7a19dd-st08z                                          3m     20Mi
-kube-system       kube-proxy-ty1m6                                                  1m     14Mi
-kube-system       spot-interrupt-handler-f7syd                                      2m     4Mi
-monitoring        prometheus-node-exporter-d91th                                    2m     9Mi
-nginx-ingress     nginx-ingress-controller-vqf8q                                    36m    151Mi
+### `top` Command
+To display the metrics of all pods on a specific node:
 ```
+$ kubenpod top nodename123
 ```
-$ kubenpod -p hoke-deployment-cda628x58-p1vzz
+To display the metrics of all pods on the node where a specific pod is running:
+```
+$ kubenpod top --pod-name mypodname
+```
 
-kubenpod ip-10-0-0-1.us-east-1.compute.internal
-hoke              hoke-deployment-cda628x58-p1vzz                                   176m   267Mi
-kube-system       aws-node-r6n1m                                                    3m     28Mi
-kube-system       coredns-165c7a19dd-st08z                                          3m     20Mi
-kube-system       kube-proxy-ty1m6                                                  1m     14Mi
-kube-system       spot-interrupt-handler-f7syd                                      2m     4Mi
-monitoring        prometheus-node-exporter-d91th                                    2m     9Mi
-nginx-ingress     nginx-ingress-controller-vqf8q                                    36m    151Mi
+### `list` Command
+To list all pods on a specific node:
 ```
+$ kubenpod list nodename123
 ```
-$ kubenpod -l ip-10-0-0-1.us-east-1.compute.internal
+To list all pods on the node where a specific pod is running:
+```
+$ kubenpod list --pod-name mypodname
+```
 
-hoke              hoke-deployment-cda628x58-p1vzz           1/1   Running     0      6d3h
-kube-system       aws-node-r6n1m                            1/1   Running     0      58d
-kube-system       coredns-165c7a19dd-st08z                  1/1   Running     0      56d
-kube-system       kube-proxy-ty1m6                          1/1   Running     0      26d
-kube-system       spot-interrupt-handler-f7syd              1/1   Running     0      5d22h
-monitoring        prometheus-node-exporter-d91th            1/1   Running     0      58d
-nginx-ingress     nginx-ingress-controller-vqf8q            1/1   Running     0      58d
-```
-```
-$ kubenpod -pl hoke-deployment-cda628x58-p1vzz
+## Additional Notes
+- Ensure that your Kubernetes context is correctly set. `kubenpod` uses the default kubeconfig file path for authentication.
+- The commands require at least cluster-level read permissions to fetch the node and pod metrics.
 
-hoke              hoke-deployment-cda628x58-p1vzz           1/1   Running     0      6d3h
-kube-system       aws-node-r6n1m                            1/1   Running     0      58d
-kube-system       coredns-165c7a19dd-st08z                  1/1   Running     0      56d
-kube-system       kube-proxy-ty1m6                          1/1   Running     0      26d
-kube-system       spot-interrupt-handler-f7syd              1/1   Running     0      5d22h
-monitoring        prometheus-node-exporter-d91th            1/1   Running     0      58d
-nginx-ingress     nginx-ingress-controller-vqf8q            1/1   Running     0      58d
-```
-```
-$ kubenpod --help
+## Contributing
+Contributions to improve Kubenpod are welcome. Please feel free to fork the repository, make changes, and submit a pull request.
 
-USAGE:
-  kubenpod <NODE_NAME>                 : show metrics of all pods of the <NODE_NAME>
-  kubenpod -p <POD_NAME>               : show metrics of pods of the <POD_NAME>'s node 
-  kubenpod                             : show this message
-  kubenpod -h,--help                   : show this message
-  kubenpod -l,--list <NODE_NAME>       : list all pods of the <NODE_NAME>
-  kubenpod -lp,--listpod <POD_NAME>    : list all pods of the <POD_NAME>'s node 
-```
+## License
+This project is licensed under the MIT License - see the LICENSE file for details.
 
 ## Special Thanks To
-
 Mert Öngengil [@mertongngl](http://github.com/mertongngl)
